server/requirements.txt
# Runtime dependencies for the OpenEnv server container.
openenv-core[core]>=0.2.2,<0.3
fastapi>=0.115.0
pydantic>=2.0
uvicorn>=0.24.0
numpy
torch
scipy
safetensors
opencv-python
mediapipe